+++ Controls/_calendar/MonthList.ts
@@ -18,7 +18,7 @@ import {IDayTemplate, IDayTemplateOptions} from 'Controls/interface';
 import {IntersectionObserverSyntheticEntry} from 'Controls/scroll';
 import {Base as dateUtils} from 'Controls/dateUtils';
 import getDimensions = require("Controls/Utils/getDimensions");
-import scrollToElement = require('Controls/Utils/scrollToElement');
+import {scrollToElement} from 'Controls/scrollUtils';
 import template = require('wml!Controls/_calendar/MonthList/MonthList');
 import monthTemplate = require('wml!Controls/_calendar/MonthList/MonthTemplate');
 import yearTemplate = require('wml!Controls/_calendar/MonthList/YearTemplate');
