+++ Controls/_input/DateTime.ts
@@ -9,9 +9,8 @@ import {
    getDefaultOptions as getValueValidatorsDefaultOptions,
    getOptionTypes as getValueValidatorsOptionTypes
 } from 'Controls/_input/interface/IValueValidators';
-import proxyModelEvents from 'Controls/Utils/proxyModelEvents';
-import {isValidDate, Container, InputContainer} from 'Controls/validate'
-import tmplNotify = require('Controls/Utils/tmplNotify');
+import {proxyModelEvents, tmplNotify} from 'Controls/eventUtils';
+import {isValidDate, Container, InputContainer} from 'Controls/validate';
 import template = require('wml!Controls/_input/DateTime/DateTime');
 
 /**
