+++ Controls/_scroll/Container.ts
@@ -21,7 +21,7 @@ import 'Controls/_scroll/Scroll/Scrollbar';
 import * as newEnv from 'Core/helpers/isNewEnvironment';
 import {SyntheticEvent} from 'Vdom/Vdom';
 import {Logger} from 'UI/Utils';
-import * as scrollToElement from 'Controls/Utils/scrollToElement';
+import {scrollToElement} from 'Controls/scrollUtils';
 import {descriptor} from 'Types/entity';
 import {detection, constants} from 'Env/Env';
 import {LocalStorageNative} from 'Browser/Storage';
