+++ Controls/_tree/TreeControl.ts
@@ -7,8 +7,7 @@ import { Map } from 'Types/shim';
 import {RecordSet} from 'Types/collection';
 
 import { saveConfig } from 'Controls/Application/SettingsController';
-import keysHandler = require('Controls/Utils/keysHandler');
-import tmplNotify = require('Controls/Utils/tmplNotify');
+import {tmplNotify, keysHandler} from 'Controls/eventUtils';
 import { MouseButtons, MouseUp } from 'Controls/Utils/MouseEventHelper';
 import { DndTreeController } from 'Controls/listDragNDrop';
 import { Controller as SourceController } from 'Controls/source';
