+++ Controls/eventUtils.ts
@@ -1,0 +1,11 @@
+/**
+ * @library Controls/eventUtils
+ * @includes tmplNotify Controls/_utils/eventUtils/ResizeObserverUtil
+ * @includes scrollToElement Controls/_utils/eventUtils/scrollToElement
+ * @public
+ * @author Красильников А.С.
+ */
+
+export {default as tmplNotify} from './_utils/eventUtils/tmplNotify';
+export {default as proxyModelEvents} from './_utils/eventUtils/proxyModelEvents';
+export {default as keysHandler} from './_utils/eventUtils/keysHandler';
