+++ Controls/scrollUtils.ts
@@ -1,0 +1,10 @@
+/**
+ * @library Controls/scrollUtils
+ * @includes ResizeObserver Controls/_utils/scrollUtils/ResizeObserverUtil
+ * @includes scrollToElement Controls/_utils/scrollUtils/scrollToElement
+ * @public
+ * @author Миронов А.Ю.
+ */
+
+export {default as ResizeObserver} from './_utils/scrollUtils/ResizeObserverUtil';
+export {default as scrollToElement} from './_utils/scrollUtils/scrollToElement';
